# Optimize query for getting tags for a level