--- a/Refresh.Interfaces.Game/Types/Levels/GameMinimalLevelResponse.cs
+++ b/Refresh.Interfaces.Game/Types/Levels/GameMinimalLevelResponse.cs
@@ -184,7 +184,7 @@ public class GameMinimalLevelResponse : IDataConvertableFrom<GameMinimalLevelRes
 
         if (dataContext.Game == TokenGame.LittleBigPlanet1)
         {
-            response.Tags = string.Join(',', dataContext.Database.GetTagsForLevel(old).Select(t => t.Tag.ToLbpString()));
+            response.Tags = string.Join(',', dataContext.Database.GetTagsForLevel(old).Select(t => t.ToLbpString()));
         }
         
         response.IconHash = dataContext.Database.GetAssetFromHash(old.IconHash)?.GetAsIcon(dataContext.Game, dataContext) ?? old.IconHash;